load csv data functionality

+++ salestrack/domain/models.py
@@ -38,7 +38,7 @@ class Sales(Base):
     id = sa.Column(sa.Integer, primary_key=True, index=True)
     product_id = sa.Column(sa.Integer, sa.ForeignKey("product.id"))
     sales_date = sa.Column(sa.Date)
-    sales_amount = sa.Column(sa.Float)
+    sales_amount = sa.Column(sa.Integer)
 
     product = sa.orm.relationship('Product')
